test: Baseball 가이드 역할을 수행하는 BaseballGuide 테스트
